sentiment-service/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
pydantic
pymongo[srv]
dnspython

python-dotenv
transformers
torch
numpy
